scripts/requirements.txt
pymysql
kafka-python
pandas 
cryptography
redis